requirements.txt
tweepy==3.10.0
python-dotenv==0.19.0

